requirements.txt
# Минимальные зависимости для VoiceMaster (clean)
aiogram>=3.4.1,<4
python-dotenv>=1.0.1
openai>=1.37.0
requests>=2.31.0
imageio-ffmpeg>=0.4.9
